+++ components/Card.js
@@ -1,49 +1,36 @@
 import Image from "next/image";
 import styled from "styled-components";
-import useLocalStorage from "use-local-storage";
+
 import HeartButton from "./HeartButton";
 import Link from "next/link";
 
-export default function Card({ plant }) {
-  const [ownedPlantIds, setOwnedPlantIds] = useLocalStorage(
-    "ownedPlantIds",
-    []
-  );
+export default function Card({ plant, handleToggleOwned }) {
   const isOwned = ownedPlantIds.includes(plant._id);
-
-  function handleToggleOwned() {
-    if (isOwned) {
-      setOwnedPlantIds(ownedPlantIds.filter((id) => id !== plant._id));
-    } else {
-      setOwnedPlantIds([...ownedPlantIds, plant._id]);
-    }
-  }
-
   return (
-      <StyledLink
-        href={`/plants/${plant._id}`}
-        aria-label={`View details for ${plant.name}`}
-      >
-        <CardWrapper>
-          <ImageWrapper>
-            <StyledImage
-              src={plant.imageUrl}
-              alt={plant.name}
-              width={300}
-              height={0}
-            />
-          </ImageWrapper>
-          <TextWrapper>
-            <HeartButton
-              isOwned={isOwned}
-              handleToggleOwned={handleToggleOwned}
-              aria-label={`Toggle owned for ${plant.name}`}
-            />
-            <h3>{plant.name}</h3>
-            <p>{plant.botanicalName}</p>
-          </TextWrapper>
-        </CardWrapper>
-      </StyledLink>
+    <StyledLink
+      href={`/plants/${plant._id}`}
+      aria-label={`View details for ${plant.name}`}
+    >
+      <CardWrapper>
+        <ImageWrapper>
+          <StyledImage
+            src={plant.imageUrl}
+            alt={plant.name}
+            width={300}
+            height={0}
+          />
+        </ImageWrapper>
+        <TextWrapper>
+          <HeartButton
+            isOwned={isOwned}
+            handleToggleOwned={handleToggleOwned}
+            aria-label={`Toggle owned for ${plant.name}`}
+          />
+          <h3>{plant.name}</h3>
+          <p>{plant.botanicalName}</p>
+        </TextWrapper>
+      </CardWrapper>
+    </StyledLink>
   );
 }
